Move all completion logic into a completion module.
