+++ setup.cfg
@@ -64,7 +64,6 @@ docs =
 	# local
 
 completion =
-	# upstream
 	shtab
 
 [options.entry_points]
